Condense schema package to 2 files
