+++ internal/schema/migrate.go
@@ -14,3 +14,70 @@ func Migrate(db *sqlx.DB) error {
 
 	return d.Migrate()
 }
+
+// migrations contains the queries needed to construct the database schema.
+// Entries should never be removed from this slice once they have been ran in
+// production.
+//
+// Using constants in a .go file is an easy way to ensure the queries are part
+// of the compiled executable and avoids pathing issues with the working
+// directory. It has the downside that it lacks syntax highlighting and may be
+// harder to read for some cases compared to using .sql files. You may also
+// consider a combined approach using a tool like packr or go-bindata.
+var migrations = []darwin.Migration{
+	{
+		Version:     1,
+		Description: "Add products",
+		Script: `
+CREATE TABLE products (
+	product_id   UUID,
+	name         TEXT,
+	cost         INT,
+	quantity     INT,
+	date_created TIMESTAMP,
+	date_updated TIMESTAMP,
+
+	PRIMARY KEY (product_id)
+);`,
+	},
+	{
+		Version:     2,
+		Description: "Add sales",
+		Script: `
+CREATE TABLE sales (
+	sale_id      UUID,
+	product_id   UUID,
+	quantity     INT,
+	paid         INT,
+	date_created TIMESTAMP,
+
+	PRIMARY KEY (sale_id),
+	FOREIGN KEY (product_id) REFERENCES products(product_id) ON DELETE CASCADE
+);`,
+	},
+	{
+		Version:     3,
+		Description: "Add users",
+		Script: `
+CREATE TABLE users (
+	user_id       UUID,
+	name          TEXT,
+	email         TEXT UNIQUE,
+	roles         TEXT[],
+	password_hash TEXT,
+
+	date_created TIMESTAMP,
+	date_updated TIMESTAMP,
+
+	PRIMARY KEY (user_id)
+);`,
+	},
+	{
+		Version:     4,
+		Description: "Add user column to products",
+		Script: `
+ALTER TABLE products
+	ADD COLUMN user_id UUID DEFAULT '00000000-0000-0000-0000-000000000000'
+`,
+	},
+}
